fix: fixed History
