+++ src/main/java/org/codegenerator/history/HistoryNode.java
@@ -1,0 +1,11 @@
+package org.codegenerator.history;
+
+import java.util.List;
+
+public interface HistoryNode<T> {
+    Object getObject();
+
+    List<HistoryCall<T>> getHistoryCalls();
+
+    HistoryType getType();
+}
