+++ src/main/java/org/codegenerator/history/HistoryType.java
@@ -1,0 +1,6 @@
+package org.codegenerator.history;
+
+public enum HistoryType {
+    ARRAY,
+    OBJECT,
+}
